requirements.txt
mysql-connector-python
nicegui
setuptools